requirements.txt
# YouTube video downloading
yt-dlp>=2024.1.0

# Audio transcription
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Optional: WhisperX with speaker diarization (requires HF token)
# whisperx>=3.7.0

# Cloud transcription services
elevenlabs>=1.0.0

# AI summarization
openai>=1.0.0
anthropic>=0.18.0

# Utility libraries
python-dotenv>=1.0.0
colorama>=0.4.6
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
